misc: fix low bat blink for real, remove mdc stub, size microopts

+++ settings.c
@@ -39,7 +39,7 @@ void SETTINGS_SaveFM(void)
 #endif
 
 	EEPROM_WriteBuffer(0x0E88, &gFM);
-	for (int i = 0; i < 5; i++) {
+	for (uint8_t i = 0; i < 5; i++) {
 		EEPROM_WriteBuffer(0x0E40 + (i * 8), &gFM_Channels[i * 4]);
 	}
 }
@@ -59,7 +59,7 @@ void SETTINGS_SaveVfoIndices(void)
 	State[3] = gEeprom.ScreenChannel[1];
 	State[4] = gEeprom.MrChannel[1];
 	State[5] = gEeprom.FreqChannel[1];
-	for (int i = 6; i < 8; i++) {
+	for (uint8_t i = 6; i < 8; i++) {
 		State[i] = 0xFF;
 	}
 
@@ -80,7 +80,7 @@ void SETTINGS_SaveSettings(void)
 	State[2] = gEeprom.TX_TIMEOUT_TIMER;
 	State[3] = 0xFF;
 	State[4] = gEeprom.KEY_LOCK;
-	for (int i = 5; i < 7; i++) {
+	for (uint8_t i = 5; i < 7; i++) {
 		State[i] = 0xFF;
 	}
 	State[7] = gEeprom.MIC_SENSITIVITY;
@@ -122,7 +122,7 @@ void SETTINGS_SaveSettings(void)
 	State[1] = gEeprom.ROGER;
 	State[2] = gEeprom.REPEATER_TAIL_TONE_ELIMINATION;
 	State[3] = gEeprom.TX_VFO;
-	for (int i = 4; i < 8; i++) {
+	for (uint8_t i = 4; i < 8; i++) {
 		State[i] = 0xFF;
 	}
 
@@ -143,7 +143,7 @@ void SETTINGS_SaveSettings(void)
 
 	State[0] = gEeprom.DTMF_CODE_PERSIST_TIME / 10U;
 	State[1] = gEeprom.DTMF_CODE_INTERVAL_TIME / 10U;
-	for (int i = 2; i < 8; i++) {
+	for (uint8_t i = 2; i < 8; i++) {
 		State[i] = 0xFF;
 	}
 
@@ -165,7 +165,7 @@ void SETTINGS_SaveSettings(void)
 	State[2] = 0xFF;
 	State[3] = gSetting_200TX;
 	State[4] = gSetting_500TX;
-	for (int i = 5; i < 8; i++) {
+	for (uint8_t i = 5; i < 8; i++) {
 		State[i] = 0xFF;
 	}
